Cambio de nombres de  ventanaUsuario a Usuario y de Usuario a ObjUsuario
